+++ src/pages/EditEntryPage.tsx
@@ -1542,7 +1542,7 @@ export default function EditEntryPage() {
       <AppShell
         sidebar={<DesktopSidebarConnected />}
         main={main}
-        bottomNav={<GlobalBottomNav authed={!!user} />}
+        bottomNav={<GlobalBottomNav authed={user !== null} />}
       />
     </>
   );